Remove usage of MockitoJUnitRunner
None of the tests in the project use Mockito's annotations, and
therefore they do not need to use MockitoJUnitRunner.

Removing it makes the code easier to maintain and refactor, and could
potentially make the tests run a bit faster (even though on my machine
the runtime improvement seems anecdotal at best).